+++ src/models/player.lua
@@ -13,9 +13,11 @@ local SPRITE_SIZE     = 8 * SPRITE_SCALE
 function player:init()
     self.maxHp = 100
     self.hp = self.maxHp
-    self.speed = 400
+    self.speed = 200
     self.velocity = vector(0, 0)
     self.sprite = tileManager.player
+    self.gunSprite = tileManager.gun
+    self.gunTip = vector(0, 0)
     self.width = SPRITE_SIZE
     self.height = SPRITE_SIZE
     self.iFrame = false
@@ -27,11 +29,10 @@ function player:init()
     self.pos = vector(x, y)
 end
 
-function player:move(dir, dt)
-    -- Normalize diagonal movements
-    if dir.x ~= 0 and dir.y ~= 0 then dir = dir:normalized() end
-
-    self.pos = self.pos + dir * self.speed * dt
+-- === Behavior ===
+function player:update(enemies, hitSound, dt)
+    self:updateIFrame(dt)
+    self:checkCollisions(enemies, hitSound)
 end
 
 -- Invincible state last for a bit
@@ -47,20 +48,56 @@ function player:updateIFrame(dt)
 end
 
 -- Check colliding with enemies
-function player:checkCollisions(enemies, dt)
+function player:checkCollisions(enemies, hitSound)
     for i, e in ipairs(enemies) do
         if not self.iFrame and collider.aabb(self, e) then
             self.hp = self.hp - e.dmg
             self.iFrame = true
+            hitSound:play()
             break
         end
     end
 end
 
--- === Behavior ===
-function player:update(enemies, dt)
-    self:updateIFrame(dt)
-    self:checkCollisions(enemies, dt)
+-- Player movements
+function player:move(dir, dt)
+    -- Normalize diagonal movements
+    if dir.x ~= 0 and dir.y ~= 0 then dir = dir:normalized() end
+
+    self.pos = self.pos + dir * self.speed * dt
+end
+
+function player:drawGun()
+    -- Center of the player
+    local cx = self.pos.x + self.width / 2
+    local cy = self.pos.y + self.height / 2
+
+    local mx, my = love.mouse:getPosition()
+    local angle = math.atan2(my - cy, mx - cx)
+    local scale = 2
+    local flip = 1
+
+    if mx < cx then
+        flip = -1
+        angle = angle + math.pi
+    end
+
+    local offset = vector(10 * flip, 0):rotated(angle)
+    local gunOrigin = vector(cx, cy) + offset
+    local gunLength = self.gunSprite:getWidth() * scale
+
+    -- Calculate tip (muzzle) of the gun
+    local muzzleDir = vector(math.cos(angle), math.sin(angle))
+    self.gunTip = gunOrigin + muzzleDir * gunLength * flip + vector(0, -4)
+
+    love.graphics.draw(
+        self.gunSprite,
+        gunOrigin.x, gunOrigin.y,
+        angle,
+        scale * flip, scale,
+        0,
+        self.gunSprite:getHeight() / 2 -- Vertically centered
+    )
 end
 
 function player:draw()
@@ -70,6 +107,8 @@ function player:draw()
         tileManager.tilemap, self.sprite, x, y,
         0, SPRITE_SCALE, SPRITE_SCALE
     )
+
+    self:drawGun()
 end
 
 function player:drawHeath()